use QProcess for running molecule editor, added file system checks

+++ adegui/generate_script.py
@@ -1,4 +1,4 @@
-import pathlib, shutil
+import pathlib, os, shutil
 
 from PyQt5.QtWidgets import QMessageBox, QFileDialog
 from adegui import Config
@@ -69,6 +69,12 @@ def write_ade_script_from_config(obj) -> None:
             if isinstance(rct_molecule.molecule, pathlib.Path):
                 xyz_fname = rct_molecule.molecule.name
                 gen_script.append(f"rct{index} = ade.Reactant('{xyz_fname}', ")
+                if not os.path.isfile(rct_molecule.molecule): # if file does not exist
+                    QMessageBox.critical(obj,
+                                         "autodE-GUI",
+                                         f"Error, {xyz_fname} file deleted or file system permission error!"
+                                         "Unable to generate script.")
+                    return None
                 shutil.copyfile(rct_molecule.molecule, cwd_path/xyz_fname)
             elif isinstance(rct_molecule.molecule, str):
                 gen_script.append(f"rct{index} = ade.Reactant(smiles='{rct_molecule.molecule}', ")
@@ -76,13 +82,20 @@ def write_ade_script_from_config(obj) -> None:
                               f"mult={rct_molecule.mult})\n")
             rct_and_prod += f"rct{index},"
 
-
     for index, prod_molecule in enumerate(Config.ade_prod_mols):
         if not prod_molecule.molecule == '':
             if isinstance(prod_molecule.molecule, pathlib.Path):
                 xyz_fname = prod_molecule.molecule.name
                 gen_script.append(f"prod{index} = ade.Product('{xyz_fname}', ")
+                if not os.path.isfile(prod_molecule.molecule):
+                    QMessageBox.critical(obj,
+                                         "autodE-GUI",
+                                         f"Error, {xyz_fname} file deleted or file system permission error!"
+                                         "Unable to generate script.")
+                    return None
                 shutil.copyfile(prod_molecule.molecule, cwd_path/xyz_fname)
+                # TODO add a check if save folder is the same (shutil crashes?)
+                # TODO refactor this part into a copy with check function
             elif isinstance(prod_molecule.molecule, str):
                 gen_script.append(f"prod{index} = ade.Product(smiles='{prod_molecule.molecule}', ")
             gen_script.append(f"charge={prod_molecule.charge}, "
